requirements.txt
absl-py==2.0.0
annotated-types==0.7.0
anyio==4.6.0
astunparse==1.6.3
bcrypt==4.2.0
cachetools==5.3.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
fonttools==4.40.0
gast==0.4.0
google-auth==2.23.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
opt-einsum==3.3.0
passlib==1.7.4
Pillow==9.5.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests-oauthlib==1.3.1
rich==13.9.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
